Avoid ambiguity or correct typo
The German word 'Gran' refers to various old measures of weight ranging from about 62 to 65 mg. The English unit grain (or troy grain), on the other hand, is a well defined unit (exactly 64.79891 mg). So we should use the word 'Grain' in German to denote the Engish unit of measurement and thus avoid any ambiguity. Wikipedia (https://de.wikipedia.org/wiki/Gran_(Einheit)#Englisches_Grain) also uses 'Grain' for the English unit.
(Not sure if Gran was chosen deliberately or if the missing i is just a typo).