Avoid confusion about symbols for atmospheres
The official symbol for the Technical atmosphere is "at", the symbol for the (standard) atmosphere (which we call physical atmosphere is German) is "atm". In the original English text and in most of the translstions, including the German one, the word atmosphere is abbreviated to Atm. This could cause confusion with the standard atmosphere whose symbol is atm. In order to avoid this possible confusion I thinks it's best to not abbreviate the word atmosphere to Atm and, to make it completely clear, add the official unit symbols.
(Translations that don't abbreviate atmosphere to Atm, like the Russian or Italian ones, do not suffer from this ambiguity).